Move getVesselControl to Mission feature
