+++ frontend/src/api/APIWorker.tsx
@@ -2,6 +2,7 @@ import { getAllBeaconMalfunctions } from '@features/BeaconMalfunction/useCases/g
 import { getVesselBeaconMalfunctions } from '@features/BeaconMalfunction/useCases/getVesselBeaconMalfunctions'
 import { openBeaconMalfunctionInKanban } from '@features/BeaconMalfunction/useCases/openBeaconMalfunctionInKanban'
 import { fleetSegmentApi } from '@features/FleetSegment/apis'
+import { getVesselControls } from '@features/Mission/useCases/getVesselControls'
 import { getAllRegulatoryLayers } from '@features/Regulation/useCases/getAllRegulatoryLayers'
 import { reportingApi } from '@features/Reporting/reportingApi'
 import { useCallback, useEffect, useRef, useState } from 'react'
@@ -14,7 +15,6 @@ import { getOperationalAlerts } from '../domain/use_cases/alert/getOperationalAl
 import { getSilencedAlerts } from '../domain/use_cases/alert/getSilencedAlerts'
 import { getAllGearCodes } from '../domain/use_cases/gearCode/getAllGearCodes'
 import { getInfractions } from '../domain/use_cases/infraction/getInfractions'
-import { getVesselControls } from '../domain/use_cases/mission/getVesselControls'
 import { getAllSpecies } from '../domain/use_cases/species/getAllSpecies'
 import { updateVesselTracks } from '../domain/use_cases/vessel/updateVesselTracks'
 import { useMainAppDispatch } from '../hooks/useMainAppDispatch'
