+++ frontend/src/features/MainWindow/components/PreviewFilteredVessels.jsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import styled from 'styled-components'
 import { useDispatch, useSelector } from 'react-redux'
-import { COLORS } from '../../constants/constants'
-import BackToVesselsListSVG from '../icons/Fleche_navigation_marees_gainsboro.svg?react'
-import { undoPreviewVessels } from '@features/Vessel/useCases/undoPreviewVessels'
+import { COLORS } from '@constants/constants.js'
+import BackToVesselsListSVG from '../../icons/Fleche_navigation_marees_gainsboro.svg?react'
+import { undoPreviewVessels } from '@features/Vessel/useCases/undoPreviewVessels.js'
 
 const PreviewFilteredVessels = () => {
   const dispatch = useDispatch()
