+++ frontend/src/features/MainWindow/index.tsx
@@ -17,7 +17,7 @@ import { ControlUnitListDialog } from '../ControlUnit/components/ControlUnitList
 import { DrawLayerModal } from '../Draw/components/DrawModal'
 import { HealthcheckHeadband } from '../Healthcheck/components/HealthcheckHeadband'
 import { LayersSidebar } from '../LayersSidebar/components'
-import PreviewFilteredVessels from '../preview_filtered_vessels/PreviewFilteredVessels'
+import PreviewFilteredVessels from './components/PreviewFilteredVessels'
 import { SideWindowLauncher } from '../SideWindow/SideWindowLauncher'
 import { VesselList } from '../Vessel/components/VesselList'
 import { VesselLoader } from '../Vessel/components/VesselLoader'
